PracticeTestsBackend/requirements.txt
pymongo
PyMuPDF
selenium
webdriver-manager
svgpathtools
PyPDF2
python-dotenv
